Factor out logic for suggesting apps and websites based on failed connections from SquidWarningService into own FailedConnectionSuggestionService.
